part4_d

+++ part4/blog/backend/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
+const { userExtractor } = require('../utils/middleware');
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -10,11 +11,14 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
   const body = request.body;
 
-  let user = await User.findById(body.userId);
-  if (!user) {
-    const users = await User.find({});
-    user = users[0];
-  }
+  const user = await User.findById(request.user.id);
+
+  // assign first user if no user
+  // let user = await User.findById(body.userId);
+  // if (!user) {
+  //   const users = await User.find({});
+  //   user = users[0];
+  // }
 
   const blog = new Blog({
     title: body.title,
@@ -37,7 +41,14 @@ blogRouter.get('/:id', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'wrong user' });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
+
   response.status(204).end();
 });
